reduce tree test redundancy
