+++ sklearn/neighbors/tests/test_ball_tree.py
@@ -3,12 +3,9 @@ import itertools
 import numpy as np
 import pytest
 from numpy.testing import assert_array_almost_equal
-from sklearn.neighbors._ball_tree import (BallTree, NeighborsHeap,
-                                          simultaneous_sort, kernel_norm,
-                                          nodeheap_sort, DTYPE, ITYPE)
+from sklearn.neighbors._ball_tree import BallTree
 from sklearn.neighbors import DistanceMetric
 from sklearn.utils import check_random_state
-from sklearn.utils._testing import assert_allclose
 
 rng = np.random.RandomState(10)
 V_mahalanobis = rng.rand(3, 3)
@@ -61,173 +58,6 @@ def test_ball_tree_query_metrics(metric):
     dist1, ind1 = bt.query(Y, k)
     dist2, ind2 = brute_force_neighbors(X, Y, k, metric)
     assert_array_almost_equal(dist1, dist2)
-
-
-def test_ball_tree_query_radius(n_samples=100, n_features=10):
-    rng = check_random_state(0)
-    X = 2 * rng.random_sample(size=(n_samples, n_features)) - 1
-    query_pt = np.zeros(n_features, dtype=float)
-
-    eps = 1E-15  # roundoff error can cause test to fail
-    bt = BallTree(X, leaf_size=5)
-    rad = np.sqrt(((X - query_pt) ** 2).sum(1))
-
-    for r in np.linspace(rad[0], rad[-1], 100):
-        ind = bt.query_radius([query_pt], r + eps)[0]
-        i = np.where(rad <= r + eps)[0]
-
-        ind.sort()
-        i.sort()
-
-        assert_array_almost_equal(i, ind)
-
-
-def test_ball_tree_query_radius_distance(n_samples=100, n_features=10):
-    rng = check_random_state(0)
-    X = 2 * rng.random_sample(size=(n_samples, n_features)) - 1
-    query_pt = np.zeros(n_features, dtype=float)
-
-    eps = 1E-15  # roundoff error can cause test to fail
-    bt = BallTree(X, leaf_size=5)
-    rad = np.sqrt(((X - query_pt) ** 2).sum(1))
-
-    for r in np.linspace(rad[0], rad[-1], 100):
-        ind, dist = bt.query_radius([query_pt], r + eps, return_distance=True)
-
-        ind = ind[0]
-        dist = dist[0]
-
-        d = np.sqrt(((query_pt - X[ind]) ** 2).sum(1))
-
-        assert_array_almost_equal(d, dist)
-
-
-def compute_kernel_slow(Y, X, kernel, h):
-    d = np.sqrt(((Y[:, None, :] - X) ** 2).sum(-1))
-    norm = kernel_norm(h, X.shape[1], kernel)
-
-    if kernel == 'gaussian':
-        return norm * np.exp(-0.5 * (d * d) / (h * h)).sum(-1)
-    elif kernel == 'tophat':
-        return norm * (d < h).sum(-1)
-    elif kernel == 'epanechnikov':
-        return norm * ((1.0 - (d * d) / (h * h)) * (d < h)).sum(-1)
-    elif kernel == 'exponential':
-        return norm * (np.exp(-d / h)).sum(-1)
-    elif kernel == 'linear':
-        return norm * ((1 - d / h) * (d < h)).sum(-1)
-    elif kernel == 'cosine':
-        return norm * (np.cos(0.5 * np.pi * d / h) * (d < h)).sum(-1)
-    else:
-        raise ValueError('kernel not recognized')
-
-
-@pytest.mark.parametrize("kernel", ['gaussian', 'tophat', 'epanechnikov',
-                                    'exponential', 'linear', 'cosine'])
-@pytest.mark.parametrize("h", [0.01, 0.1, 1])
-@pytest.mark.parametrize("rtol", [0, 1E-5])
-@pytest.mark.parametrize("atol", [1E-6, 1E-2])
-@pytest.mark.parametrize("breadth_first", [True, False])
-def test_ball_tree_kde(kernel, h, rtol, atol, breadth_first, n_samples=100,
-                       n_features=3):
-    rng = np.random.RandomState(0)
-    X = rng.random_sample((n_samples, n_features))
-    Y = rng.random_sample((n_samples, n_features))
-    bt = BallTree(X, leaf_size=10)
-
-    dens_true = compute_kernel_slow(Y, X, kernel, h)
-
-    dens = bt.kernel_density(Y, h, atol=atol, rtol=rtol,
-                             kernel=kernel,
-                             breadth_first=breadth_first)
-    assert_allclose(dens, dens_true,
-                    atol=atol, rtol=max(rtol, 1e-7))
-
-
-def test_gaussian_kde(n_samples=1000):
-    # Compare gaussian KDE results to scipy.stats.gaussian_kde
-    from scipy.stats import gaussian_kde
-    rng = check_random_state(0)
-    x_in = rng.normal(0, 1, n_samples)
-    x_out = np.linspace(-5, 5, 30)
-
-    for h in [0.01, 0.1, 1]:
-        bt = BallTree(x_in[:, None])
-        gkde = gaussian_kde(x_in, bw_method=h / np.std(x_in))
-
-        dens_bt = bt.kernel_density(x_out[:, None], h) / n_samples
-        dens_gkde = gkde.evaluate(x_out)
-
-        assert_array_almost_equal(dens_bt, dens_gkde, decimal=3)
-
-
-def test_ball_tree_two_point(n_samples=100, n_features=3):
-    rng = check_random_state(0)
-    X = rng.random_sample((n_samples, n_features))
-    Y = rng.random_sample((n_samples, n_features))
-    r = np.linspace(0, 1, 10)
-    bt = BallTree(X, leaf_size=10)
-
-    D = DistanceMetric.get_metric("euclidean").pairwise(Y, X)
-    counts_true = [(D <= ri).sum() for ri in r]
-
-    def check_two_point(r, dualtree):
-        counts = bt.two_point_correlation(Y, r=r, dualtree=dualtree)
-        assert_array_almost_equal(counts, counts_true)
-
-    for dualtree in (True, False):
-        check_two_point(r, dualtree)
-
-
-
-
-def test_neighbors_heap(n_pts=5, n_nbrs=10):
-    heap = NeighborsHeap(n_pts, n_nbrs)
-
-    for row in range(n_pts):
-        d_in = rng.random_sample(2 * n_nbrs).astype(DTYPE, copy=False)
-        i_in = np.arange(2 * n_nbrs, dtype=ITYPE)
-        for d, i in zip(d_in, i_in):
-            heap.push(row, d, i)
-
-        ind = np.argsort(d_in)
-        d_in = d_in[ind]
-        i_in = i_in[ind]
-
-        d_heap, i_heap = heap.get_arrays(sort=True)
-
-        assert_array_almost_equal(d_in[:n_nbrs], d_heap[row])
-        assert_array_almost_equal(i_in[:n_nbrs], i_heap[row])
-
-
-def test_node_heap(n_nodes=50):
-    vals = rng.random_sample(n_nodes).astype(DTYPE, copy=False)
-
-    i1 = np.argsort(vals)
-    vals2, i2 = nodeheap_sort(vals)
-
-    assert_array_almost_equal(i1, i2)
-    assert_array_almost_equal(vals[i1], vals2)
-
-
-def test_simultaneous_sort(n_rows=10, n_pts=201):
-    dist = rng.random_sample((n_rows, n_pts)).astype(DTYPE, copy=False)
-    ind = (np.arange(n_pts) + np.zeros((n_rows, 1))).astype(ITYPE, copy=False)
-
-    dist2 = dist.copy()
-    ind2 = ind.copy()
-
-    # simultaneous sort rows using function
-    simultaneous_sort(dist, ind)
-
-    # simultaneous sort rows using numpy
-    i = np.argsort(dist2, axis=1)
-    row_ind = np.arange(n_rows)[:, None]
-    dist2 = dist2[row_ind, i]
-    ind2 = ind2[row_ind, i]
-
-    assert_array_almost_equal(dist, dist2)
-    assert_array_almost_equal(ind, ind2)
 
 
 def test_query_haversine():
